move search tab to 'search_tab' folder
